Revert "1. try fixing user name chaos"
This reverts commit be64daa359592d0e69a488f91568da4c00ef3704.

The main point of this "chaos" is, that a user has its username that is provided by telegram. (The profile username in Telegram). The User should be able to choose a "calling-name" or "chosen-username" for each group individually. Therefore the chosen-username field is only present in the GroupUserAcoount Class and not the User class.